need to fix slow sample rate

+++ src/Communicator.h
@@ -17,23 +17,22 @@
 
 	class Communicator{
 	public:
-			// delete these methods to prevent making accidental copies of the class.
-			Communicator(Communicator const &) = delete;
-			void operator=(Communicator const &) = delete;
+    // delete these methods to prevent making accidental copies of the class.
+    Communicator(Communicator const &) = delete;
+    void operator=(Communicator const &) = delete;
 
-			static Communicator* getInstance();
+    static Communicator* getInstance();
 
-			void sendDataPoint(int32_t displacement, float load, uint8_t measurementStage);
-			void sendDataEnd();
-			char getCommand();
-			void sendCommand(char command, int16_t data);
-			void sendCommand(char command);
-			int16_t getInt();
-			MeasurementParams receiveMeasurementParams();
+    void sendDataPoint(int32_t displacement, int16_t load, uint8_t measurementStage);
+    void sendDataEnd();
+    char getCommand();
+    void sendCommand(char command, int16_t data);
+    void sendCommand(char command);
+    int16_t getInt();
+    MeasurementParams receiveMeasurementParams();
 
 
 	private:
-			Communicator();
-
+		Communicator();
 	};
 #endif
